More type errors from auto convertin the json

+++ src/main/java/io/mindspice/jxch/rpc/schemas/wallet/WalletBalance.java
@@ -5,15 +5,15 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public record WalletBalance(
 
 	@JsonProperty("wallet_id") int walletId,
-	@JsonProperty("pending_change") int pendingChange,
+	@JsonProperty("pending_change") long pendingChange,
 	@JsonProperty("unspent_coin_count") int unspentCoinCount,
 	@JsonProperty("fingerprint") long fingerprint,
 	@JsonProperty("wallet_type") int walletType,
-	@JsonProperty("confirmed_wallet_balance") int confirmedWalletBalance,
+	@JsonProperty("confirmed_wallet_balance") long confirmedWalletBalance,
 	@JsonProperty("pending_coin_removal_count") int pendingCoinRemovalCount,
 	@JsonProperty("asset_id") String assetId,
-	@JsonProperty("spendable_balance") int spendableBalance,
-	@JsonProperty("max_send_amount") int maxSendAmount,
-	@JsonProperty("unconfirmed_wallet_balance") int unconfirmedWalletBalance
+	@JsonProperty("spendable_balance") long spendableBalance,
+	@JsonProperty("max_send_amount") long maxSendAmount,
+	@JsonProperty("unconfirmed_wallet_balance") long unconfirmedWalletBalance
 ) {
 }
